Added vocatives, mild deprecation
"Not self-depreciation, deprecation. They're two separate things."

Turns out I didn't notice the missing vocatives until too late - it should be fixed now. It uncovered a minor bug, which should now be patched.

Grammar checker updated to not allow pi phrases to be added to be alone or attached to things *they should not be attached to*